perf(test): split daemon.test.ts across three files, halving the suite's floor
