+++ cli/.changelog/next/RUSH-3230.md
@@ -1,0 +1,8 @@
+- **Split `daemon.test.ts`, the suite's floor (RUSH-3230).** It was 34 subprocess tests in
+  one file at ~53s locally (159s on a loaded worker) — the slowest file in the repo, and
+  therefore the whole suite's floor, because vitest parallelises across *files* and runs
+  one file's tests sequentially in a single worker. Now three files by theme (command
+  surface / services + webhooks / doctor + logs) over a shared
+  `daemon-test-harness.ts`, so they run concurrently: **53.5s → 26s wall, all 35 tests
+  still passing**. No test was changed, skipped, or dropped. Source:
+  `cli/src/commands/daemon-test-harness.ts`.
